requirements.txt
# Core Libraries
tensorflow>=2.12.0
numpy>=1.21.6
pandas>=1.3.0
scipy>=1.7.0

# Visualization Libraries
matplotlib>=3.4.0
seaborn>=0.11.2

# Data Management and Experiment Tracking
dvc>=2.20.0
mlflow==2.2.2
gdown>=4.5.0

# Utilities
python-box==6.0.2
PyYAML>=6.0
tqdm>=4.64.0
ensure==1.0.2
joblib>=1.2.0
types-PyYAML>=6.0.0
pyarrow

# Flask for API Development
Flask>=2.2.0
Flask-Cors>=3.0.0

# Notebook Environment
notebook>=6.5.0

# Editable Installation for Local Package
-e .
